added cast to double
[From SVN r2217 (trunk)]

+++ tests/tget_d.c
@@ -43,7 +43,7 @@ check_denorms ()
       rnd_mode = GMP_RNDN;
       for (k = -17; k <= 17; k += 2)
         {
-          d = k * DBL_MIN; /* k * 2^(-1022) */
+          d = (double) k * DBL_MIN; /* k * 2^(-1022) */
           f = 1.0;
           mpfr_set_si (x, k, GMP_RNDN);
           mpfr_div_2exp (x, x, 1022, GMP_RNDN); /* k * 2^(-1022) */
